Moved tests into appropriate packages
