+++ src/test/java/lithium/university/TwitterCacheTest.java
@@ -52,7 +52,7 @@ public class TwitterCacheTest {
         assertNotNull(twitterCacheTest.getCachedList());
 
         try {
-            Thread.sleep(ttlSeconds * 1000);
+            Thread.sleep((ttlSeconds + 1)* 1000);
         } catch (Exception e) {
             fail();
         }
